skills/octo/scripts/requirements.txt
requests>=2.28.0
jsonschema>=4.17.0
pyyaml>=6.0
